Big modifs: enhance script by using generic player initialisation:
When opening json stats, call an init players function that check if the player has all the stats that the script takes into account. If not create them.
When a new player enter in the game, the create it fully as usualy.
Then do not need anymore to check the except KeyError exception.
For special cases as the 'kill'-'player'-'victimName' where the victimName may not exist yet, then use the addDictIfNotExist function to create it.